part3

+++ firstgame/src/controllers/update-character.controller.ts
@@ -17,10 +17,9 @@ import {
   requestBody,
   HttpErrors,
 } from '@loopback/rest';
-import {Character,IdSet, Armor, Weapon, Skill} from '../models';
+import {Character, Armor, Weapon, Skill} from '../models';
 import {
   CharacterRepository,
-  IdSetRepository,
   WeaponRepository,
   ArmorRepository,
   SkillRepository
@@ -33,8 +32,6 @@ export class UpdateCharacterController {
     public characterRepository : CharacterRepository,
 
     //add following lines
-    @repository(IdSetRepository)
-    public idSetRepository : IdSetRepository,
     @repository(WeaponRepository)
     public weaponRepository : CharacterRepository,
     @repository(ArmorRepository)
@@ -59,13 +56,18 @@ export class UpdateCharacterController {
     @param.path.string('id') id: string,
   ): Promise<any[]> {
     let res: any[] = ['no weapon', 'no armor', 'no skill'];
-    //todo
-    try{res[0] = await this.characterRepository.weapon(id).get()}
-    catch{console.log('@get /updatecharacter/{id}: no current weapon')};
-    try{res[1] = await this.characterRepository.armor(id).get()}
-    catch{console.log('@get /updatecharacter/{id}: no current armor')};
-    try{res[2] = await this.characterRepository.skill(id).get()}
-    catch{console.log('@get /updatecharacter/{id}: no current skill')};
+
+    let filter: Filter = {where:{"characterId":id}};
+    if((await this.weaponRepository.find(filter))[0] != undefined){
+      res[0] = await this.characterRepository.weapon(id).get()
+    }
+    if((await this.armorRepository.find(filter))[0] != undefined){
+      res[1] = await this.characterRepository.armor(id).get()
+    }
+    if((await this.skillRepository.find(filter))[0] != undefined){
+      res[2] = await this.characterRepository.skill(id).get()
+    }
+
     return res;
   }
 
@@ -131,15 +133,6 @@ export class UpdateCharacterController {
       await this.characterRepository.weapon(id).delete();
     }
     await this.characterRepository.updateById(id, char);
-
-    //generate weapon ID
-    let weaponId: string = uuid();
-    while(await this.idSetRepository.exists(weaponId)){
-      weaponId = uuid();
-    }
-    weapon.id = weaponId;
-    let wId : Partial<IdSet> = {"id":weaponId};
-    await this.idSetRepository.create(wId);
     return await this.characterRepository.weapon(id).create(weapon);
   }
 
@@ -160,29 +153,21 @@ export class UpdateCharacterController {
     @param.path.string('id') id: string,
     @requestBody() armor: Armor,
   ): Promise<Armor> {
+    //equip new armor
     let char: Character = await this.characterRepository.findById(id);
     char.attack! += armor.attack;
     char.defence! += armor.defence;
-    try{
+
+    //unequip old armor
+    let filter: Filter = {where:{"characterId":id}};
+    if((await this.armorRepository.find(filter))[0] != undefined){
       let oldArmor: Armor = await this.characterRepository.armor(id).get();
       char.attack! -= oldArmor.attack!;
       char.defence! -= oldArmor.defence!;
       await this.characterRepository.armor(id).delete();
     }
-    catch(e){
-      console.log('@patch /updatecharacter/{id}/armor: no current armor');
-    }
-    finally{
-      let armorId: string = uuid();
-      while(await this.idSetRepository.exists(armorId)){
-        armorId = uuid();
-      }
-      armor.id = armorId;
-      let aId : Partial<IdSet> = {};
-      aId.id = armorId;
-      await this.idSetRepository.create(aId);
-      return await this.characterRepository.armor(id).create(armor);
-    }
+    await this.characterRepository.updateById(id, char);
+    return await this.characterRepository.armor(id).create(armor);
   }
 
   /**
@@ -202,23 +187,13 @@ export class UpdateCharacterController {
     @param.path.string('id') id: string,
     @requestBody() skill: Skill,
   ): Promise<Skill> {
-    try{//todo
+
+    //unequip old armor
+    let filter: Filter = {where:{"characterId":id}};
+    if((await this.skillRepository.find(filter))[0] != undefined){
       await this.characterRepository.skill(id).delete();
     }
-    catch(e){
-      console.log('@patch /updatecharacter/{id}/skill: no current skill');
-    }
-    finally{
-      let skillId: string = uuid();
-      while(await this.idSetRepository.exists(skillId)){
-        skillId = uuid();
-      }
-      skill.id = skillId;
-      let sId : Partial<IdSet> = {};
-      sId.id = skillId;
-      await this.idSetRepository.create(sId);
-      return await this.characterRepository.skill(id).create(skill);
-    }
+    return await this.characterRepository.skill(id).create(skill);
   }
 
   /**
@@ -235,16 +210,15 @@ export class UpdateCharacterController {
   async deleteWeapon(
     @param.path.string('id') id: string
   ): Promise<void> {
-    try{
+    //unequip old weapon
+    let filter: Filter = {where:{"characterId":id}};
+    if((await this.weaponRepository.find(filter))[0] != undefined){
       let oldWeapon: Weapon = await this.characterRepository.weapon(id).get();
-      await this.characterRepository.weapon(id).delete();
       let char: Character = await this.characterRepository.findById(id);
       char.attack! -= oldWeapon.attack!;
       char.defence! -= oldWeapon.defence!;
+      await this.characterRepository.weapon(id).delete();
       await this.characterRepository.updateById(id, char);
-    }
-    catch(e){
-      console.log('@del /updatecharacter/weapon: no current weapon');
     }
   }
 
@@ -262,16 +236,15 @@ export class UpdateCharacterController {
   async deleteArmor(
     @param.path.string('id') id: string
   ): Promise<void> {
-    try{
+    //unequip old armor
+    let filter: Filter = {where:{"characterId":id}};
+    if((await this.armorRepository.find(filter))[0] != undefined){
       let oldArmor: Armor = await this.characterRepository.armor(id).get();
-      await this.characterRepository.armor(id).delete();
       let char: Character = await this.characterRepository.findById(id);
       char.attack! -= oldArmor.attack!;
       char.defence! -= oldArmor.defence!;
+      await this.characterRepository.armor(id).delete();
       await this.characterRepository.updateById(id, char);
-    }
-    catch(e){
-      console.log('@del /updatecharacter/armor: no current armor');
     }
   }
